Updated XML File: web.xml
Added <servlet>
This basically is like a class path, the serlvets Class Name, and the
path to the package its located in.

Added <servlet-mapping>
Class Name
The URL patter, this response to JSP File /register.jsp so /register is
the URL pattern.

This has to be done for each additional  Servlets.