Updated Data Definition File: FWRPSchema.sql
Was having errors inserting CharitableOrganiztion as a user type.

updated column in table user:

UserType ENUM('Retailer', 'Consumer', 'CharitableOrganization') NOT NULL

To  UserType ENUM('RETAILER', 'CONSUMER', 'CHARITABLE_ORGANIZATION') NOT
NULL

Now the values are exactly the same in java and in the SQL database,
this resolved the insertion error for Charitable Organization user type.