Updated Java Servlet File: InventoryServlet.java
After 3-4 hours of dealing with pathing error, infinite loops on load,
only loading at one URL but not the other, not loading the page at all
but calling the servlet.

Its finally working and updating the
inventory.jsp page correctly with the information in the database.

I have no idea why something this simple ended up being so hard to get
working correctly