Updated Java Servlet File: AddItemToInventoryServlet.java
Finally was able to get this working correctly.

Added Logic to handle updating inventory items.

Updated url paths to work correctly when called.

doGet() handles response to button pressed in Retailer User Dashboard,
checks to make sure valid user and is logged in.

doPost() handles the heavy lifting of adding items to the database.

Took some trial and error but this servlet now lets the user add items
to the inventory through the web app, specifically the Retailer User
Dashboard